misc: Fix format truncation in ssh_path_expand_escape()
error: ‘%u’ directive output may be truncated writing between 1 and 10
bytes into a region of size 6.

(cherry picked from commit 20406e51c9e1e096dc8ba47975abad448a51bfc1)

+++ include/libssh/session.h
@@ -223,7 +223,7 @@ struct ssh_session_struct {
         char *agent_socket;
         unsigned long timeout; /* seconds */
         unsigned long timeout_usec;
-        unsigned int port;
+        uint16_t port;
         socket_t fd;
         int StrictHostKeyChecking;
         char compressionlevel;
